requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=3.0.0
paho-mqtt>=1.6.1
PyMySQL>=1.0.2
APScheduler>=3.10.0
psycopg2-binary>=2.9.9
